+++ language-plutus-napkin/bench/Bench.hs
@@ -4,7 +4,7 @@ module Main (main) where
 
 import           Criterion.Main
 import qualified Data.ByteString.Lazy  as BSL
-import           Language.PlutusNapkin
+import           Language.PlutusCore
 
 tinyProgram :: BSL.ByteString
 tinyProgram = "[(builtin addInteger) x y]"
