+++ language-plutus-napkin/src/Language/PlutusCore/PrettyPrint.hs
@@ -1,10 +1,10 @@
-module Language.PlutusNapkin.PrettyPrint ( prettyPrint
+module Language.PlutusCore.PrettyPrint ( prettyPrint
                                          ) where
 
 import qualified Data.ByteString.Lazy             as BSL
 import qualified Data.IntMap                      as IM
-import           Language.PlutusNapkin.Identifier
-import           Language.PlutusNapkin.Type
+import           Language.PlutusCore.Identifier
+import           Language.PlutusCore.Type
 
 prettyPrint :: (IdentifierState, Term a) -> Maybe BSL.ByteString
 prettyPrint (is, Var _ (Name _ (Unique i))) = IM.lookup i (fst is)
